BoundingBoxAttachmentModel updated
