+++ Tests/SpineTests/Resources/skins.json
@@ -14,6 +14,13 @@
                     "width": 51,
                     "height": 19
                 }
+            },
+            "head-bb": {
+                "head": {
+                    "type": "boundingbox",
+                    "vertexCount": 6,
+                    "vertices": [-19.14, -70.3, 40.8, -118.07, 257.77, -115.62, 285.16, 57.18, 120.77, 164.95, -5.07, 76.95]
+                }
             }
         }
     },
